fix(review): pre-merge fixes from the dev->main review (crc deadlock + 7 more) (#216)
Found by the reviewing-alp-changes pass over the dev->main delta (13 check_*.py
gates already green; these are what they can't catch). Adversarially verified.

CRITICAL:
- crc_alif.c: crc_update() returned -ENOTSUP (non-multiple-of-4 for CRC32) WITHOUT
  releasing the begin()-held semaphore -> the CRC engine deadlocked permanently
  for every later crc_begin(). Now gives the lock + resets state on that error.

CORRECTNESS:
- counter_alif_lptimer.c: cancel_alarm() did not clear data->running (set_alarm
  sets it) -> a later start() was wrongly rejected -EALREADY. Clear it.
- lprtc_calendar_shim.c: elapsed_s * freq_hz used a 32-bit product that overflows
  near the one-CCVR-wrap boundary (129600 s * 32768 Hz > 4.2e9), corrupting the
  sub-second remainder. Use a 64-bit intermediate.
- firmware/cc3501e/protocol.c: a handler over-reporting data_len would truncate
  silently into the uint16 reply length. Clamp to the reply cap + fail closed.
- adc/alif_e8.c: documented that set_trigger_source records but does not yet apply
  an external trigger (software single-shot read path) -- no longer a silent no-op.

DOCTRINE:
- can_cast.c + sdhc_dwc.c: added the missing ADR-0017 Tier-2 status headers.
- aen-ethernet-link: cited the bench-observed ETH_CTRL 0x4903F080 provenance.
- E1M-AEN701.yaml: alif_mhu -> alif_mhuv2 to match the confirmed AEN801 token.

Verified: crc + lptimer still RESULT PASS on E8 (Flow C); adc/rtc examples build;
cc3501e protocol.c syntax-clean. Deferred (need design/HW, tracked): the CAN
z_remove_filter cb_table slot reuse (already documented, planned v0.3 map) and the
ADC HW-trigger wiring.

+++ src/backends/rtc/lprtc_calendar_shim.c
@@ -216,7 +216,10 @@ static alp_status_t shim_get_time(alp_rtc_backend_state_t *st, alp_rtc_time_t *t
 	/* Unsigned 32-bit wrap gives the true elapsed ticks across one CCVR wrap. */
 	uint32_t elapsed_ticks = now - s->tick_snapshot;
 	int64_t  elapsed_s     = (int64_t)elapsed_ticks / (int64_t)s->freq_hz;
-	uint32_t rem_ticks     = elapsed_ticks - (uint32_t)elapsed_s * s->freq_hz;
+	/* 64-bit intermediate: elapsed_s * freq_hz can exceed UINT32_MAX near the
+	 * one-CCVR-wrap boundary (129600 s * 32768 Hz > 4.2e9), so a 32-bit product
+	 * would wrap and corrupt rem_ticks. The remainder (< freq_hz) still fits u32. */
+	uint32_t rem_ticks = elapsed_ticks - (uint32_t)(elapsed_s * (int64_t)s->freq_hz);
 
 	int64_t unix_s = s->epoch_base + elapsed_s;
 	_unix_to_calendar(unix_s, t);
